+++ src-docs/src/views/page/page_header.js
@@ -21,7 +21,7 @@ export default () => (
       },
     ]}
     description="This description should be describing the current page as depicted by the page title. It has the grow prop set to false on the EuiText block so that it is the proper line length. And it will also never extend beneath the right side content (buttons)."
-    rightSideContent={[
+    rightSideItems={[
       <EuiButton fill>Add something</EuiButton>,
       <EuiButton>Do something</EuiButton>,
     ]}>